# Implement `foldMap1` for `NonEmpty` in terms of `foldr`